+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+# Version 0
+
+### 0.1.0
+- Add attachment related models
+- Add helper to add attachments to a typescript content type
+- Add helper functions to create apis for single content types
+
 ### 0.0.0
 - Initial release